[flutter_tools] Refactor DeferredComponents validator and add build target (#76400)
